requirements.txt
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
